requirements.txt
fastapi
uvicorn
celery==5.3.5
redis==5.0.2
ocrmypdf
pillow